Move the displays of data to a new file so the main js files are not populated
